Replace GenServer types with :gen_statem ones where needed (#21)
The GenServer types seem to be the same types, but they could easily diverge at a later time, so it's better to use the `:gen_statem` types. The only GenServer type left is `GenServer.options`, where it's actually what the functions expect and it's different to what is passed to `:gen_statem`.